[platform/broadcom]: Makefile changes for arista-fwutil
#### Description for the changelog

Add facilities to include arista fwutil into image if deb is in repo. Arista fwutil is proprietary and will be included in private repos

<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
This utility is required for updating optics firmware on eos->sonic conversion for chassis

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
Add makefile entries to copy and install the deb if it exists

#### How to verify it
Build image with the deb in platform/broadcom/extra-debs and see if it is in the image
<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ x] 202405

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->