[Moby] Update interface name and alias for back panel ports, and lanemap and polarity settings for loopback connections for LAB HWSKU
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
There are some issues with the existing Moby platform files

##### Interface names and alias for interfaces in the back panel ports

Based on our internal discussion with Microsoft, update the interface name and alias so these backpanel ports can be properly numbered and labelled

##### Lanemap and polarity settings of the back panel ports

In order to use a loopback setup between the back panel ports, e.g. by using `LAB` HWSKU, we need to update the lanemap and polarity of these ASIC PHY cores accordingly

#### How I did it

Update the corresponding platform files based on the discussion with Microsoft

#### How to verify it

Use latest image on Moby device and get links up

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [x] msft-202412

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [x] msft-202412

#### A picture of a cute animal (not mandatory but encouraged)

![41b086054ba785e8f728a93aa5d267a0](https://github.com/user-attachments/assets/4d22e537-4558-4152-b630-10324fa94bea)